Move the core dynamic tools functionality to drmcpbase to be used by global mcp
