+++ pyproject.toml
@@ -4,7 +4,7 @@ build-backend = "setuptools.build_meta"
 
 [project]
 name = "datarobot-genai"
-version = "0.15.123"
+version = "0.15.124"
 description = "Generic helpers for GenAI"
 readme = "README.md"
 requires-python = ">=3.11, <3.14"
@@ -43,6 +43,7 @@ where = ["src"]
 datarobot_genai = [
     "py.typed",
     "drmcp/core/dynamic_tools/deployment/schemas/*.json",
+    "drmcpbase/dynamic_tools/deployment/schemas/*.json",
 ]
 
 [tool.uv]
