+++ src/datarobot_genai/drmcp/core/enums.py
@@ -15,25 +15,6 @@ from enum import Enum
 from enum import auto
 
 
-class DataRobotMCPToolCategory(Enum):
-    USER_TOOL = auto()  # tools created by users
-    BUILT_IN_TOOL = auto()  # tools as a wrapper of external service
-    USER_TOOL_DEPLOYMENT = auto()  # tools dynamically loaded after MCP server is up
-
-    @staticmethod
-    def from_string(enum_str: str) -> "DataRobotMCPToolCategory":
-        enum_str_map = {
-            "USER_TOOL": DataRobotMCPToolCategory.USER_TOOL,
-            "BUILT_IN_TOOL": DataRobotMCPToolCategory.BUILT_IN_TOOL,
-            "USER_TOOL_DEPLOYMENT": DataRobotMCPToolCategory.USER_TOOL_DEPLOYMENT,
-        }
-        if enum_str not in enum_str_map:
-            error_msg = f"Enum string should be one of {', '.join(enum_str_map.keys())}"
-            raise ValueError(error_msg)
-
-        return enum_str_map[enum_str]
-
-
 class DataRobotMCPPromptCategory(Enum):
     USER_PROMPT_TEMPLATE = auto()  # prompt created by users
     USER_PROMPT_TEMPLATE_VERSION = auto()  # prompts dynamically loaded after MCP server is up
