+++ src/datarobot_genai/drmcpbase/dynamic_tools/deployment/schemas/drum_prediction_fallback_schema.json
