+++ tests/drmcp/unit/test_dr_mcp_server.py
@@ -21,9 +21,9 @@ import pytest
 from fastmcp import FastMCP
 
 from datarobot_genai.drmcp.core.dr_mcp_server import DataRobotMCPServer
-from datarobot_genai.drmcp.core.dynamic_tools.deployment.adapters.default import Metadata
 from datarobot_genai.drmcp.core.mcp_instance import DataRobotMCP
 from datarobot_genai.drmcp.core.mcp_instance import mcp
+from datarobot_genai.drmcpbase.dynamic_tools.deployment.adapters.default import Metadata
 
 
 @pytest.fixture
